Extract the rules into a separate file.
